* Addition of new Feeds (PeersFeed / TransferFeed / MsgFeed)  * Addition of new FeedItems (MsgItem / ChanNewItem )  * Removed CheckBoxes at the top of NewsFeed (should be in config)  * Enabled subscribe/unsubscribeButtons in ChannelFeed.  * Enabled ChanNewItem and MsgItem in NewsFeeds.  * Remove Goto Section Button from FeedItems.  * Disabled PlayMedia button - if no attachments.  * Enabled Drag from Search Window (with new class SearchTreeWidget)  * Enabled Drag from SharedFiles Dialog (mods to RemoteDirModel).  * Enabled Drop in GeneralMsgDialog from Search/SharedFiles.  * Updated Rs Interface (64 bits for filesize)  * Other bits and bobs.
git-svn-id: http://svn.code.sf.net/p/retroshare/code/trunk@635 b45a01b8-16f6-495d-af2f-9b41ad6348cc

+++ retroshare-gui/src/rsiface/rstypes.h
@@ -201,7 +201,7 @@ class DirDetails
 	std::string name;
 	std::string hash;
 	std::string path;
-	uint32_t count;
+	uint64_t count;
 	uint32_t age;
 	uint32_t rank;
 
@@ -215,7 +215,7 @@ class FileDetail
 	std::string name;
 	std::string hash;
 	std::string path;
-	uint32_t size;
+	uint64_t size;
 	uint32_t age;
 	uint32_t rank;
 };
